Restructured input handling, replaced camera with a top-down camera

+++ GameFrame/src/NG/ActionHandling/MouseScrollListener.java
@@ -4,5 +4,5 @@ package NG.ActionHandling;
  * @author Geert van Ieperen. Created on 20-9-2018.
  */
 public interface MouseScrollListener {
-    void mouseScrolled(double value);
+    void mouseScrolled(float value);
 }
